+++ example/lib/module/pages.dart
@@ -40,8 +40,6 @@ class _ImagePageState extends State<ImagePage> with TickerProviderStateMixin {
   }
 }
 
-
-
 class ToastPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -159,6 +157,67 @@ class ExpansionTilesPage extends StatelessWidget {
                   margin: const EdgeInsets.all(12),
                   alignment: Alignment.centerLeft,
                   child: BasisText('item$index', color: Colors.black)))),
+        ]);
+  }
+}
+
+class SimpleBuilderPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    const int i = 0;
+    return OverlayScaffold(
+        backgroundColor: Colors.white,
+        appBar:
+            AppBar(title: const Text('SimpleBuilder Demo'), centerTitle: true),
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[
+          ValueBuilder<int>(
+              initialValue: i,
+              builder: (BuildContext context, int value,
+                  ValueCallback<int> updater) {
+                return Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      IconBox(
+                          icon: Icons.remove_circle_outline,
+                          onTap: () {
+                            updater(value -= 1);
+                          }),
+                      Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: Text(value.toString()),
+                      ),
+                      IconBox(
+                          icon: Icons.add_circle_outline,
+                          onTap: () {
+                            updater(value += 1);
+                          })
+                    ]);
+              }),
+          const Center(child: SizedBox(height: 30)),
+          ValueListenBuilder<int>(
+              initialValue: 1,
+              builder:
+                  (BuildContext context, ValueNotifier<int> valueListenable) {
+                return Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      IconBox(
+                          icon: Icons.remove_circle_outline,
+                          onTap: () {
+                            valueListenable.value -= 1;
+                          }),
+                      Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: Text(valueListenable.value.toString()),
+                      ),
+                      IconBox(
+                          icon: Icons.add_circle_outline,
+                          onTap: () {
+                            valueListenable.value += 1;
+                          })
+                    ]);
+              }),
         ]);
   }
 }