+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [1.9.10]
+  * add [ValueListenBuilder] and Example
+  * [ValueBuilder] add BuildContext for builder and add Example
 ## [1.9.9]
   * fix bug
 ## [1.9.7]